refactor(update): isolate service refresh environment helpers
